Move properties to sub folder
