+++ src/docx/properties/properties.ts
@@ -1,5 +1,5 @@
-import { DocumentAttributes } from "../docx/document/document-attributes";
-import { XmlComponent } from "../docx/xml-components";
+import { DocumentAttributes } from "../document/document-attributes";
+import { XmlComponent } from "../xml-components";
 import { Created, Creator, Description, Keywords, LastModifiedBy, Modified, Revision, Subject, Title } from "./components";
 
 export interface IPropertiesOptions {
